fixes and added example suites
